requirements.txt
django-celery-beat == 2.6.0
pytz == 2024.1
python-decouple == 3.8
redis == 5.0.5
python-dotenv == 1.0.1

